Убрала extend, убрала map в mistakes, переименовала тест utils, экспорт App на уровне компонента
