+++ src/utils.js
@@ -1,7 +1,3 @@
 export const capitalizeFirstLetter = (word) => {
   return `${word.charAt(0).toUpperCase()}${word.slice(1)}`;
 };
-
-export const extend = (a, b) => {
-  return Object.assign({}, a, b);
-};
